git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@3543 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/gui/components/WbTable.java
@@ -359,7 +359,7 @@ public class WbTable
 		};
 		configureEnterKeyAction(a);
 	}
-	
+
 	public void configureEnterKeyAction(Action enterAction)
 	{
 		this.getInputMap(WHEN_FOCUSED).put(WbSwingUtilities.ENTER, "wbtable-stop-editing");
@@ -2284,7 +2284,7 @@ public class WbTable
 		}
 	}
 
-	public long addRow()
+	public int addRow()
 	{
 		DataStoreTableModel ds = this.getDataStoreTableModel();
 		if (ds == null) return -1;
